+++ docs/changelog.rst
@@ -7,6 +7,7 @@ Changelog
 * restore Python 3.7 compatibility
 * add default behavior when `Page.input` has a `url` attribute.
 * add :py:class:`PdfPage`
+* add `page_to_items` helper
 * add `Page.example_input` and `Page.example_source` for test command
 * add `Page.logger` for logging
 * add `Workflow.yield_items` method